requirements.txt
stravalib
requests
geopy
python-dotenv